cctv-backend-content/requirements.txt
psycopg2
requests
python-dotenv
numpy
scikit-learn
readerwriterlock
